Create Arduino indoor DHT sensor
Arduino sketch for DHT22/AM2302 indoor temperature/humidity sensor using the MySensors library. Also uses NRF24L01 radio module.  On this mesh network, this sensor is Node 3. Child 0 is humidity, Child 2 is temperature. These values are sent to an Arduino Mega with an ethernet shield. The Mega acts as a bridge to the larger network.