Create Arduino Outdoor DHT sensor
This Arduino sketch is used for an outdoor DHT22/AM2302 temperature and humidity sensor read by an Arduino Nano. This sensor is Node 2, humidity is Child 0 and temperature is Child 1.  Data protocols are set by MySensors (see MySensors.org). The Nano sends the data through a mesh network using a NRF24L01 radio module. The data is picked up by an Arduino Mega 2560 that acts as a network bridge. It sends the data through ethernet to a router, and is picked up by a Raspberry Pi acting as a MQTT broker.