Create Arduino Mega Bridge
This Arduino sketch is run on an Arduino Mega 2560 with an ethernet shield and an NRF24L01 radio module. The Mega accepts data packets from other Arduinos and their associated devices. It acts as a gateway to send MQTT messages to a Raspberry Pi.  The Pi runs Openhab2 and acts as an MQTT broker. The sketch gives the Mega an IP address ending in 107. The router sets the Mega as 107 through DCHP statically. See MySensors.org and Openhab.org for documentation.